Ignore textcritical † character
From Victoria: "The second one is an easier fix. † before and after the word is not a Greek character, but it is a textcritical one telling the reader that this word is not properly attested in this place. Can we somewhat just take all the † out of the text? That would fix a large number of errors and they are really only a textcritical tool. "